Implemented Feature v0.4- Authentication

+++ src/features/actions/deleteFlights.ts
@@ -8,17 +8,16 @@ import type { UUID } from "@/src/schemas/flightSchemas";
 export async function deleteFlightById(flightId: UUID) {
 
     const supabase = await createClient();
-    /*const {data: {user} } = await supabase.auth.getUser();
+
+    const {data: {user} } = await supabase.auth.getUser();
     
     if (!user) throw new Error("Unauthorised!");
-    */
-
 
     const { error } = await supabase
         .from('flights')
         .delete()
-        .eq('id', flightId);
-        // .eq('user_id', user.id);
+        .eq('id', flightId)
+        .eq('user_id', user.id);
 
     if (error) throw new Error(error.message)
